+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 ## [0.1.8] - 2026-02-11
 
-* Added new `timeSync` error with factory method.
+* Added new `invalidTime` error with factory method.
 
 ## [0.1.7] - 2024-05-23